trying ensemble

+++ scripts/workers/run-bowtie2-build.sh
@@ -2,9 +2,7 @@
 
 #PBS -W group_list=bhurwitz
 #PBS -q standard
-#PBS -l jobtype=cluster_only
-#PBS -l select=1:ncpus=12:mem=23gb:pcmem=2gb
-#PBS -l pvmem=46gb
+#PBS -l select=1:ncpus=12:mem=23gb
 #PBS -l walltime=24:00:00
 #PBS -l cput=24:00:00
 #PBS -M scottdaniel@email.arizona.edu
@@ -20,8 +18,10 @@ echo "Bowtie2 indexing..."
 
 cd $(dirname $MOUSEBT2)
 
-bowtie2-build --large-index $MOUSEFASTA $MOUSEBT2
+bowtie2-build $MOUSEFASTA $MOUSEBT2
 
-ln $MOUSEFASTA "$MOUSEBT2".fa
+if [[ ! -e "$MOUSEBT2".fa ]]; then
+    ln $MOUSEFASTA "$MOUSEBT2".fa
+fi
 
 echo "Done $(date)"
